requirements.txt
flask
python-dotenv
gunicorn
flask_sqlalchemy
psycopg2-binary
cloudinary
Werkzeug
flask-migrate